requirements.txt
scipy==1.8.1
pandas==1.5.1
gym==0.21.0
stable_baselines3==1.6.2
seaborn==0.13.0
numpy==1.24.4
casadi==3.6.4
pyyaml==6.0.1
matplotlib==3.7.4
pytest==7.2.0
control_ablations @ git+https://github.com/racheltrimble/control_ablations_oc_paper@78dd03b